Semester 3/DJANGO-SWE_PROJECT/Back-End-Part/HOSPITAL_MANAGEMENT/BACK-END/contact_us/serializers.py
from rest_framework import serializers
from contact_us.models import ContactUs

class Contact_UsSerializer(serializers.ModelSerializer):
    class Meta:
        model=ContactUs
        fields='__all__'