Semester 3/DJANGO-SWE_PROJECT/Back-End-Part/HOSPITAL_MANAGEMENT/BACK-END/requirements.txt
asgiref
Django
django-environ
django-filter
djangorestframework
Markdown
pillow
sqlparse
tzdata
